+++ AvgScript/src/webview/_onClickLink.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import { jumpToDocument } from '../lib/utilities';
 
 export const onClickLinkScript = `
 <script>
@@ -20,32 +21,6 @@ export async function handleOnClickLink(message: any) {
         const uri = vscode.Uri.parse(message.link);
         const line = parseInt(uri.fragment);
 
-        if (Number.isNaN(line)) {
-            const editor = await vscode.window.showTextDocument(uri
-                , {
-                    viewColumn: vscode.ViewColumn.Beside
-                });
-            editor.revealRange(new vscode.Range(0, 0, 0, 0)
-                , vscode.TextEditorRevealType.InCenterIfOutsideViewport);
-        }
-
-        const doc = await vscode.workspace.openTextDocument(uri);
-        const text = doc.lineAt(line).text;
-
-        const range = new vscode.Range(line, 0
-            , line, text.length);
-        const selection = new vscode.Selection(line, 0
-            , line, text.length);
-
-        const editor = await vscode.window.showTextDocument(uri, {
-            viewColumn: vscode.ViewColumn.Beside,
-            // viewColumn: vscode.window.activeTextEditor === undefined || vscode.window.activeTextEditor.viewColumn === undefined
-            //     ? vscode.ViewColumn.Beside
-            //     : vscode.window.activeTextEditor.viewColumn + 1,
-            selection: selection,
-        });
-
-        editor.revealRange(range
-            , vscode.TextEditorRevealType.InCenterIfOutsideViewport);
+        jumpToDocument(uri, line);
     }
 }